requirements.txt
flask>=3.1.0
gunicorn>=23.0.0
PyMuPDF>=1.25.0
google-genai>=1.0.0
python-docx>=1.1.0
mammoth>=1.8.0
pydantic>=2.0.0
